Change request attribute set for amazon auth
This is a fix for a change in the way that boto signs requests, introduced in 2.7.0
The commit hash causing this change is 4fb722fe5780f97d7b97d45e36c382e79eca315c

We previously set an attribute called `path` on our tornado requests because boto
used it to sign the request. Tornado does not use this attribute for anything. Boto
now looks for `auth_path` instead, so I am switching accordingly